+++ packages/adapter-mysql/src/index.ts
@@ -1,2 +1,2 @@
-export { mysql2Adapter as mysql2 } from "./mysql2/index.js";
-export { planetscaleAdapter as planetscale } from "./planetscale/index.js";
+export { mysql2Adapter as mysql2 } from "./drivers/mysql2.js";
+export { planetscaleAdapter as planetscale } from "./drivers/planetscale.js";