+++ packages/adapter-postgresql/src/utils.ts
@@ -1,33 +1,29 @@
-import type { SessionSchema, UserSchema, KeySchema } from "lucia-auth";
-import { ColumnValue } from "./query.js";
-
-export const transformDatabaseSession = (
-	session: PostgresSessionSchema
-): SessionSchema => {
-	return {
-		id: session.id,
-		user_id: session.user_id,
-		active_expires: Number(session.active_expires),
-		idle_expires: Number(session.idle_expires)
+const createPreparedStatementHelper = (
+	placeholder: (index: number) => string
+) => {
+	const helper = (
+		values: Record<string, any>
+	): readonly [fields: string[], placeholders: string[], arguments: any[]] => {
+		const keys = Object.keys(values);
+		return [
+			keys.map((k) => escapeName(k)),
+			keys.map((_, i) => placeholder(i)),
+			keys.map((k) => values[k])
+		] as const;
 	};
+	return helper;
 };
 
-export const transformDatabaseKey = (key: PostgresKeySchema): KeySchema => {
-	return {
-		id: key.id,
-		user_id: key.user_id,
-		primary_key: Boolean(key.primary_key),
-		hashed_password: key.hashed_password,
-		expires: key.expires === null ? null : Number(key.expires)
-	};
+const ESCAPE_CHAR = `"`;
+
+export const escapeName = (val: string) => {
+	return `${ESCAPE_CHAR}${val}${ESCAPE_CHAR}`;
 };
 
-type PgSchema<_Schema extends Record<string, ColumnValue>> = {
-	[K in keyof _Schema]: Extract<_Schema[K], number> extends never
-		? _Schema[K]
-		: _Schema[K] | string;
-};
+export const helper = createPreparedStatementHelper((i) => `$${i + 1}`);
 
-export type PostgresSessionSchema = PgSchema<SessionSchema>;
-export type PostgresKeySchema = PgSchema<KeySchema>;
-export type PostgresUserSchema = PgSchema<UserSchema>;
+export const getSetArgs = (fields: string[], placeholders: string[]) => {
+	return fields
+		.map((field, i) => [field, placeholders[i]].join(" = "))
+		.join(",");
+};