+++ packages/integration-oauth/src/utils.ts
@@ -1,3 +1,0 @@
-export type AwaitedReturnType<T extends (...args: any[]) => any> = Awaited<
-	ReturnType<T>
->;