+++ packages/adapter-mysql/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @lucia-auth/adapter-mysql
 
+## 2.0.0-beta.4
+
+### Major changes
+
+- [#788](https://github.com/pilcrowOnPaper/lucia/pull/790) by [@pilcrowOnPaper](https://github.com/pilcrowOnPaper) : Require `lucia@2.0.0-beta.3`
+
+## 2.0.0-beta.3
+
+### Patch changes
+
+- [#768](https://github.com/pilcrowOnPaper/lucia/pull/768) by [@pilcrowOnPaper](https://github.com/pilcrowOnPaper) : Update peer dependency
+
+## 2.0.0-beta.2
+
+### Patch changes
+
+- [#756](https://github.com/pilcrowOnPaper/lucia/pull/756) by [@pilcrowOnPaper](https://github.com/pilcrowOnPaper) : Fix peer dependency version
+
+## 2.0.0-beta.1
+
+- [#731](https://github.com/pilcrowOnPaper/lucia/pull/731) by [@pilcrowOnPaper](https://github.com/pilcrowOnPaper) : Fix internal imports
+
+### Patch changes
+
+- [#729](https://github.com/pilcrowOnPaper/lucia/pull/729) by [@pilcrowOnPaper](https://github.com/pilcrowOnPaper) : Export `d1` adapter
+
+## 2.0.0-beta.0
+
+### Major changes
+
+- [#682](https://github.com/pilcrowOnPaper/lucia/pull/682) by [@pilcrowOnPaper](https://github.com/pilcrowOnPaper) : Require `lucia@^2.0.0`
+
+### Minor changes
+
+- [#682](https://github.com/pilcrowOnPaper/lucia/pull/682) by [@pilcrowOnPaper](https://github.com/pilcrowOnPaper) : Add option to configure table names in `mysql2()` and `planetscale()`
+
 ## 1.1.1
 
 ### Patch changes