+++ packages/adapter-prisma/CHANGELOG.md
@@ -1,10 +1,43 @@
 # @lucia-auth/adapter-prisma
 
+## 3.0.0-beta.3
+
+### Major changes
+
+- [#788](https://github.com/pilcrowOnPaper/lucia/pull/790) by [@pilcrowOnPaper](https://github.com/pilcrowOnPaper) : Require `lucia@2.0.0-beta.3`
+
+## 3.0.0-beta.2
+
+### Patch changes
+
+- [#768](https://github.com/pilcrowOnPaper/lucia/pull/768) by [@pilcrowOnPaper](https://github.com/pilcrowOnPaper) : Update peer dependency
+
+## 3.0.0-beta.1
+
+### Major changes
+
+- [#755](https://github.com/pilcrowOnPaper/lucia/pull/755) by [@pilcrowOnPaper](https://github.com/pilcrowOnPaper) : Update `prisma()` parameters
+
+### Patch changes
+
+- [#756](https://github.com/pilcrowOnPaper/lucia/pull/756) by [@pilcrowOnPaper](https://github.com/pilcrowOnPaper) : Fix peer dependency version
+
+## 3.0.0-beta.0
+
+### Major changes
+
+- [#682](https://github.com/pilcrowOnPaper/lucia/pull/682) by [@pilcrowOnPaper](https://github.com/pilcrowOnPaper) : Require `lucia@^2.0.0`
+
+  - Export adapter as named exports (`prisma()`)
+
+  - Update adapter params
+
 ## 2.0.0
 
 ### Major changes
 
 - [#529](https://github.com/pilcrowOnPaper/lucia/pull/529) by [@pilcrowOnPaper](https://github.com/pilcrowOnPaper) : [Breaking] Requires `lucia-auth@^1.3.0`
+
   - Update to new specifications
 
 ## 1.0.0